requirements.txt
# 基础依赖
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
transformers>=4.12.0
datasets>=1.15.0

# Jupyter和开发工具
jupyter>=1.0.0
ipywidgets>=7.6.0
tqdm>=4.62.0
wandb>=0.12.0

# 文件处理和可视化
pillow>=8.3.0
opencv-python>=4.5.0
plotly>=5.3.0

# Colab连接工具
pyngrok>=5.1.0
google-colab-ssh>=0.1.0
paramiko>=2.8.0

# 其他工具
requests>=2.26.0
pathlib2>=2.3.6
python-dotenv>=0.19.0 